requirements.txt
aniso8601==9.0.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
checksumdir==1.2.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
passlib==1.7.4
pytz==2023.3
requests==2.31.0
resources==0.0.1
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
urllib3==2.0.4
Werkzeug==2.2.3
WTForms==3.0.1
